Rename StiffnessParticle -> Node (there was no more difference between them)
